requirements.txt
pytest~=8.3.5
h11~=0.14.0
pip~=23.2.1
attrs~=25.3.0
PySocks~=1.7.1
cffi~=1.17.1
typing_extensions~=4.12.2
trio~=0.29.0
outcome~=1.3.0.post0
sniffio~=1.3.1
sortedcontainers~=2.4.0
idna~=3.10
pluggy~=1.5.0
iniconfig~=2.0.0
certifi~=2025.1.31
urllib3~=2.3.0
charset-normalizer~=3.4.1
selenium~=4.29.0
trio-websocket~=0.12.2
websocket-client~=1.8.0
allure-python-commons~=2.13.5
packaging~=24.2
pycparser~=2.22
wsproto~=1.2.0
requests~=2.32.3
webdriver-manager~=4.0.2
python-dotenv~=1.0.1